grafik jadi tinggal koneksi realtime ke data base dan penambahan fitur lainnya
