+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Cobalagi"
         tools:targetApi="31">
         <activity
-            android:name=".grafik_data"
+            android:name=".Grafik_data"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
