+++ app/src/main/java/com/example/cobalagi/MainActivity.java
@@ -4,7 +4,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -86,7 +85,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void onCLikgoToGrafik (View view) {
-        Intent intent = new Intent(this, grafik_data.class);
+        Intent intent = new Intent(this, Grafik_data.class);
         startActivity(intent);
     }
